Přidaní ERROR LOGS + BE,FE, přetypování všech server actions - typova kontrola

+++ src/services/posts.ts
@@ -1,4 +1,5 @@
 import { deleteImageFromCloudinary } from "@/app/actions"
+import { handleError } from "@/helpers/handleError"
 import { IPost } from "@/interfaces/interfaces"
 import { getPublicIdFromUrl } from "@/lib/cloudinary"
 import { connectDB } from "@/lib/mongo"
@@ -6,13 +7,17 @@ import { Post } from "@/models/Post"
 import { Reactions } from "@/models/Reactions"
 import { auth } from "@clerk/nextjs/server"
 
+export type UpdatePostServiceType =
+  | { success: true }
+  | { success: false, errMsg: string }
+
 export async function updatePostService(newPost: IPost) {
     
     const { userId } = await auth()
     const { _id, userID, ...updateData } = newPost
 
     if(userId !== newPost.userID) {
-        return { success: false, errMsg: "You cannot do that" }
+        throw new Error("You cannot do that")
     }
 
     try {
@@ -23,38 +28,46 @@ export async function updatePostService(newPost: IPost) {
             { new: true }
         )
         if (!result) {
-            return { success: false, errMsg: "Somethng went wrong" }
+            throw new Error("Somethng went wrong")
         }
-        return { success: true }
+
+        return { success: true } as const
     } catch (error: unknown) {
-        if (error instanceof Error) {
-            return { success: false, errMsg: error.message }
-        }
-        return { success: false, errMsg: "Unknown error" }
+        const { errMsg } = await handleError(error, {
+            action: "Update Post",
+            component: "updatePostService()",
+            input: { newPost, userId}
+        })
+        return { success: false, errMsg }
     }
 }
 
-export async function deletePostService(postID: string) {
+export type DeletePostServiceType =
+  | { success: true }
+  | { success: false, errMsg: string }
+
+export async function deletePostService(postID: string) : Promise<DeletePostServiceType> {
 
     const { userId } = await auth()
+    if(!userId) throw new Error("Unauthorized")
 
     try {
         await connectDB()
+        
         const post = await Post.findById(postID)
-
         if (!post) {
-            return { success: false, errMsg: "Post not found" }
+            throw new Error("Post not found")
         }
 
         if (post.userID !== userId) {
-            return { success: false, errMsg: "You cannot do that" }
+            throw new Error("You cannot do that")
         }
 
         if (post.images.length > 0) {
             for (const x of post.images) {
                 const res = await deleteImageFromCloudinary(x)
                 if(!res.success) {
-                    return { success: false, errMsg: "Error during deleting image from cloud" }
+                    throw new Error("Error during deleting image from cloud")
                 }
             }
         }
@@ -65,15 +78,18 @@ export async function deletePostService(postID: string) {
         return { success: true }
 
     } catch (error: unknown) {
-        if (error instanceof Error) {
-            return { success: false, errMsg: error.message }
-        }
-        return { success: false, errMsg: "Unknown error" }
+        const { errMsg } = await handleError(error, {
+            action: "Delete Post",
+            component: "deletePostService()",
+            input: { postID, userId}
+        })
+        return { success: false, errMsg }
     }
 }
 
 export async function deleteAllPostsService(userID: string) {
     try {
+        await connectDB()
         const posts = await Post.find({ userID })
         if(posts.length === 0) return { success: true }
